add new funcion library for interrupt uses

+++ getDHT22.h
@@ -1,15 +1,13 @@
 #include "DHT.h"
-
 #define DHTPIN A5
-
 #define DHTTYPE DHT22
 
-int tStatus = 0;
-int hStatus = 0;
+uint8_t  tStatus = 0;
+uint8_t  hStatus = 0;
 
 DHT dht(DHTPIN, DHTTYPE);
 
-int gettStatus(float t) {
+uint8_t gettStatus(float t) {
   if (t > 18.0 && t < 22.0) {
     tStatus = 0;
   } else  if ((t > 15.0 && t < 18.0) || (t > 22 && t < 30)) {
@@ -21,7 +19,7 @@ int gettStatus(float t) {
   return tStatus;
 }
 
-int gethStatus(int h) {
+uint8_t gethStatus(int h) {
   if (h > 40 && h < 60) {
     hStatus = 0;
   } else  if ((h > 30 && h < 75) || (h > 75 && h < 80)) {
